# docs/openapi.yaml
openapi: 3.0.3
info:
  title: 兜行 Douxing API
  description: |
    兜行平台 REST API（`packages/server`）。

    - 统一响应：`{ code, message, messageKey?, data }`，`code === 0` 为成功
    - 鉴权：Header `Authorization: Bearer <Access Token JWT>`；过期后客户端用 Refresh Token 调 `POST /api/auth/refresh`（见 [双Token认证与无感刷新.md](./双Token认证与无感刷新.md)）
    - 国际化：Header `Accept-Language: zh-CN | en-US`
    - 人类可读说明见 [API接口文档.md](./API接口文档.md)
  version: 1.4.1
  contact:
    name: 兜行项目
servers:
  - url: http://localhost:3000
    description: 本地开发
  - url: https://api.yxtao.site
    description: 生产环境
...
security:
  - bearerAuth: []
paths:
...
  /api/attractions:
    get:
      tags: [景点]
      summary: 景点列表（已发布）
      security: []
      operationId: attractionsList
      parameters:
        - name: city
          in: query
          schema: { type: string }
        - name: cityCode
          in: query
          schema: { type: string }
        - name: category
          in: query
          schema:
            type: string
            enum: [attraction, restaurant, hotel]
        - name: keyword
          in: query
          schema: { type: string }
        - name: tags
          in: query
          schema: { type: string }
          description: 逗号分隔标签
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100 }
        - name: offset
          in: query
          schema: { type: integer, minimum: 0 }
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
  /api/attractions/{id}:
    get:
      tags: [景点]
      summary: 景点详情
      security: []
      operationId: attractionsDetail
      parameters:
        - $ref: '#/components/parameters/AttractionId'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/attractions/admin/pending:
    get:
      tags: [景点]
      summary: 待审核景点（管理员）
      operationId: attractionsAdminPending
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedApiResponse'
  /api/attractions/admin/catalog:
    get:
      tags: [景点]
      summary: 景点管理列表（管理员）
      operationId: attractionsAdminCatalog
      parameters:
        - name: city
          in: query
          schema: { type: string }
        - name: keyword
          in: query
          schema: { type: string }
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedApiResponse'
  /api/attractions/admin/{id}/cover:
    post:
      tags: [景点]
      summary: 上传景点封面（管理员）
      operationId: attractionsAdminCover
      parameters:
        - $ref: '#/components/parameters/AttractionId'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: 图片，≤ 3MB
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/attractions/admin/{id}/cover/refresh-amap:
    post:
      tags: [景点]
      summary: 高德刷新封面（管理员）
      operationId: attractionsAdminCoverRefreshAmap
      parameters:
        - $ref: '#/components/parameters/AttractionId'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/attractions/admin/{id}/approve:
    post:
      tags: [景点]
      summary: 审核通过景点（管理员）
      operationId: attractionsAdminApprove
      parameters:
        - $ref: '#/components/parameters/AttractionId'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/attractions/admin/media/pending:
    get:
      tags: [景点]
      summary: 待审核路线视频列表（管理员 · H10-b）
      operationId: attractionsAdminMediaPending
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedApiResponse'
  /api/attractions/admin/media/{mediaId}/review:
    patch:
      tags: [景点]
      summary: 审核路线视频（管理员 · H10-b）
      operationId: attractionsAdminMediaReview
      parameters:
        - name: mediaId
          in: path
          required: true
          schema: { type: integer, minimum: 1 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [approved, rejected]
                rejectReason:
                  type: string
                  maxLength: 256
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
  /api/routes:
    get:
      tags: [路线]
      summary: 路线列表
      operationId: routesList
      parameters:
        - name: scope
          in: query
          schema:
            type: string
            enum: [mine, hot, favorites, plaza]
        - name: status
          in: query
          schema: { type: integer, enum: [0, 1, 2] }
          description: 0 草稿 / 1 已发布 / 2 归档
        - name: sort
          in: query
          schema:
            type: string
            enum: [recent, hot, views]
        - name: all
          in: query
          schema: { type: string, enum: ['1'] }
          description: 管理员查看全站
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: limit
          in: query
          schema: { type: integer }
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedApiResponse'
...
  /api/routes/{id}:
    get:
      tags: [路线]
      summary: 路线详情
      operationId: routesDetail
      parameters:
        - $ref: '#/components/parameters/RouteId'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    put:
      tags: [路线]
      summary: 编辑草稿路线
      operationId: routesUpdateDraft
      parameters:
        - $ref: '#/components/parameters/RouteId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRouteDraftRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
  /api/share/routes/{id}:
    get:
      tags: [分享]
      summary: 公开路线只读摘要
      security: []
      operationId: shareRouteDetail
      parameters:
        - $ref: '#/components/parameters/RouteId'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "登录后获得的 Access Token（JWT），填入 Authorization: Bearer <token>"
  parameters:
...
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    RouteId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
    AttractionId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
...
  schemas:
    ApiResponse:
      type: object
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: 成功
        messageKey:
          type: string
          example: api.ok
        data:
          nullable: true
    PaginatedApiResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/PaginatedResult'
    PaginatedResult:
      type: object
      properties:
        items:
          type: array
          items: {}
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
        hasMore:
          type: boolean
...
    UpdateRouteDraftRequest:
      type: object
      properties:
        name: { type: string }
        description: { type: string, nullable: true }
        budgetRange: { type: string, nullable: true }
        days: { type: integer }
        interestTags:
          type: array
          items: { type: string }
        routeDetail:
          type: object
...